- Moved request access form to its own component - Pulled form elements out of Home component
