+++ tdrs-frontend/src/components/Home/Home.jsx
@@ -8,6 +8,7 @@ import FormGroup from '../FormGroup'
 import STTComboBox from '../STTComboBox'
 import { requestAccess } from '../../actions/requestAccess'
 import ResourceCards from '../ResourceCards'
+import RequestAccessForm from '../RequestAccessForm/RequestAccessForm'
 import {
   accountStatusIsApproved,
   accountIsInReview,
@@ -20,110 +21,11 @@ import {
  * permissions once they are approved by an Admin in the backend.
  */
 function Home() {
-  const errorRef = useRef(null)
-
   const user = useSelector((state) => state.auth.user)
-
-  const [errors, setErrors] = useState({})
-  const [profileInfo, setProfileInfo] = useState({
-    firstName: '',
-    lastName: '',
-    stt: '',
-  })
-  const dispatch = useDispatch()
-  const [touched, setTouched] = useState({})
-
   const sttList = useSelector((state) => state?.stts?.sttList)
 
   const userAccessInReview = useSelector(accountIsInReview)
   const userAccessRequestApproved = useSelector(accountStatusIsApproved)
-
-  const [jurisdictionType, setJurisdictionTypeInputValue] = useState('state')
-  const shouldShowSttComboBox = !user?.email?.includes('@acf.hhs.gov')
-
-  const validation = (fieldName, fieldValue) => {
-    const field = {
-      firstName: 'First Name',
-      lastName: 'Last Name',
-      stt: shouldShowSttComboBox && 'A state, tribe, or territory',
-    }[fieldName]
-
-    if (
-      Boolean(field) &&
-      typeof fieldValue === 'string' &&
-      fieldValue.trim() === ''
-    ) {
-      return `${field} is required`
-    }
-    return null
-  }
-
-  const setJurisdictionType = (val) => {
-    setStt('')
-    setJurisdictionTypeInputValue(val)
-  }
-
-  const setStt = (sttName) => {
-    setProfileInfo((currentState) => ({
-      ...currentState,
-      stt: sttName,
-    }))
-  }
-
-  const handleChange = ({ name, value }) => {
-    setProfileInfo({ ...profileInfo, [name]: value })
-  }
-
-  const handleBlur = (evt) => {
-    const { name, value } = evt.target
-
-    const { [name]: removedError, ...rest } = errors
-
-    const error = validation(name, value)
-
-    setErrors({
-      ...rest,
-      ...(error && { [name]: touched[name] && error }),
-    })
-  }
-
-  const handleSubmit = (evt) => {
-    evt.preventDefault()
-
-    // validate the form
-    const formValidation = Object.keys(profileInfo).reduce(
-      (acc, key) => {
-        const newError = validation(key, profileInfo[key])
-        const newTouched = { [key]: true }
-        return {
-          errors: {
-            ...acc.errors,
-            ...(newError && { [key]: newError }),
-          },
-          touched: {
-            ...acc.touched,
-            ...newTouched,
-          },
-        }
-      },
-      {
-        errors: { ...errors },
-        touched: { ...touched },
-      }
-    )
-    setErrors(formValidation.errors)
-    setTouched(formValidation.touched)
-
-    if (!Object.values(formValidation.errors).length) {
-      return dispatch(
-        requestAccess({
-          ...profileInfo,
-          stt: sttList.find((stt) => stt.name === profileInfo.stt),
-        })
-      )
-    }
-    return setTimeout(() => errorRef.current.focus(), 0)
-  }
 
   if (userAccessInReview) {
     return (
@@ -141,112 +43,7 @@ function Home() {
       </div>
     )
   } else if (!userAccessRequestApproved) {
-    return (
-      <div className="margin-top-5 margin-bottom-5">
-        <p className="margin-top-1 margin-bottom-4">
-          Please enter your information to request access from an OFA
-          administrator
-        </p>
-        <p>Fields marked with an asterisk (*) are required.</p>
-        <form className="usa-form" onSubmit={handleSubmit}>
-          <div
-            className={`usa-error-message ${
-              !!Object.keys(errors).length && !!Object.keys(touched).length
-                ? 'display-block'
-                : 'display-none'
-            }`}
-            ref={errorRef}
-            tabIndex="-1"
-            role="alert"
-          >
-            There are {Object.keys(errors).length} errors in this form
-          </div>
-          <FormGroup
-            error={errors.firstName}
-            name="firstName"
-            label="First Name"
-            inputValue={profileInfo.firstName}
-            handleChange={handleChange}
-            handleBlur={handleBlur}
-          />
-          <FormGroup
-            error={errors.lastName}
-            name="lastName"
-            label="Last Name"
-            inputValue={profileInfo.lastName}
-            handleChange={handleChange}
-            handleBlur={handleBlur}
-          />
-          {shouldShowSttComboBox && (
-            <div className="usa-form-group">
-              <fieldset className="usa-fieldset">
-                <legend className="usa-label text-bold">
-                  Jurisdiction Type
-                </legend>
-                <div className="usa-radio">
-                  <input
-                    className="usa-radio__input"
-                    id="state"
-                    type="radio"
-                    name="jurisdictionType"
-                    value="state"
-                    defaultChecked
-                    onChange={() => setJurisdictionType('state')}
-                  />
-                  <label className="usa-radio__label" htmlFor="state">
-                    State
-                  </label>
-                </div>
-                <div className="usa-radio">
-                  <input
-                    className="usa-radio__input"
-                    id="tribe"
-                    type="radio"
-                    name="jurisdictionType"
-                    value="tribe"
-                    onChange={() => setJurisdictionType('tribe')}
-                  />
-                  <label className="usa-radio__label" htmlFor="tribe">
-                    Tribe
-                  </label>
-                </div>
-                <div className="usa-radio">
-                  <input
-                    className="usa-radio__input"
-                    id="territory"
-                    type="radio"
-                    name="jurisdictionType"
-                    value="territory"
-                    onChange={() => setJurisdictionType('territory')}
-                  />
-                  <label className="usa-radio__label" htmlFor="territory">
-                    Territory
-                  </label>
-                </div>
-              </fieldset>
-            </div>
-          )}
-          {jurisdictionType && shouldShowSttComboBox && (
-            <div
-              className={`usa-form-group ${
-                errors.stt ? 'usa-form-group--error' : ''
-              }`}
-            >
-              <STTComboBox
-                selectStt={setStt}
-                error={Boolean(errors.stt)}
-                selectedStt={profileInfo?.stt}
-                handleBlur={handleBlur}
-                sttType={jurisdictionType}
-              />
-            </div>
-          )}
-          <Button type="submit" className="width-full request-access-button">
-            Request Access
-          </Button>
-        </form>
-      </div>
-    )
+    return <RequestAccessForm user={user} sttList={sttList} />
   }
 
   return (
